+++ storybook/cypress/support/image-snapshot-voorbeeldpaginas.ts
@@ -1,10 +1,24 @@
 import { waitForComponents } from "./wait-for-components";
 
-export function voorbeeldpaginaImageSnapshots(label: string, voorbeeldpaginas: string[], args?: string) {
+interface Viewport {
+  width: number;
+  height: number;
+}
+
+interface ExamplePage {
+  id: string;
+  viewport?: Viewport;
+}
+
+export function voorbeeldpaginaImageSnapshots(label: string, voorbeeldpaginas: ExamplePage[], args?: string) {
   describe(`Voorbeeldpagina image snapshots (${label})`, () => {
-    for (const id of voorbeeldpaginas) {
-      it(`matches image snapshot of ${id} (${label})`, () => {
-        cy.visit(`http://localhost:45000/iframe.html?id=${id}${args ? `&args=${args}` : ""}`);
+    for (const voorbeelpagina of voorbeeldpaginas) {
+      it(`matches image snapshot of ${voorbeelpagina.id} (${label})`, () => {
+        if (voorbeelpagina.viewport) {
+          cy.viewport(voorbeelpagina.viewport.width, voorbeelpagina.viewport.height);
+        }
+
+        cy.visit(`http://localhost:45000/iframe.html?id=${voorbeelpagina.id}${args ? `&args=${args}` : ""}`);
 
         waitForComponents();